first commit at partial objs
