+++ discord/iterators.py
@@ -957,8 +957,8 @@ class EntitlementIterator(_AsyncIterator["Entitlement"]):
         state,
         user_id: int | None = None,
         sku_ids: list[int] | None = None,
-        before: datetime.datetime | Object | None = None,
-        after: datetime.datetime | Object | None = None,
+        before: datetime.datetime | Snowflake | None = None,
+        after: datetime.datetime | Snowflake | None = None,
         limit: int | None = None,
         guild_id: int | None = None,
         exclude_ended: bool | None = None,
