+++ discord/user.py
@@ -36,9 +36,11 @@ from .flags import PublicUserFlags
 from .iterators import EntitlementIterator
 from .monetization import Entitlement
 from .utils import MISSING, Undefined, _bytes_to_base64_data, snowflake_time
+from .partials import PartialUser
 
 if TYPE_CHECKING:
     from datetime import datetime
+    from typing_extensions import Self
 
     from .abc import Snowflake, SnowflakeTime
     from .channel import DMChannel
@@ -46,7 +48,6 @@ if TYPE_CHECKING:
     from .message import Message
     from .app.state import ConnectionState
     from .types.channel import DMChannel as DMChannelPayload
-    from .types.user import PartialUser as PartialUserPayload
     from .types.user import User as UserPayload
 
 
@@ -63,13 +64,8 @@ class _UserTag:
     id: int
 
 
-class BaseUser(_UserTag):
+class BaseUser(_UserTag, PartialUser):
     __slots__ = (
-        "name",
-        "id",
-        "discriminator",
-        "global_name",
-        "_avatar",
         "_banner",
         "_accent_colour",
         "bot",
@@ -80,22 +76,13 @@ class BaseUser(_UserTag):
     )
 
     if TYPE_CHECKING:
-        name: str
-        id: int
-        discriminator: str
-        global_name: str | None
         bot: bool
         system: bool
         _state: ConnectionState
-        _avatar: str | None
         _banner: str | None
         _accent_colour: int | None
         _avatar_decoration: dict | None
         _public_flags: int
-
-    def __init__(self, *, state: ConnectionState, data: UserPayload | PartialUserPayload) -> None:
-        self._state = state
-        self._update(data)
 
     def __repr__(self) -> str:
         if self.is_migrated:
@@ -122,21 +109,15 @@ class BaseUser(_UserTag):
     def __eq__(self, other: Any) -> bool:
         return isinstance(other, _UserTag) and other.id == self.id
 
-    def __hash__(self) -> int:
-        return self.id >> 22
-
-    async def _update(self, data: UserPayload) -> None:
-        self.name = data["username"]
-        self.id = int(data["id"])
-        self.discriminator = data["discriminator"]
-        self.global_name = data.get("global_name", None) or None
-        self._avatar = data["avatar"]
+    async def _update(self, data: UserPayload) -> Self:
+        await super()._update(data)
         self._banner = data.get("banner", None)
         self._accent_colour = data.get("accent_color", None)
         self._avatar_decoration = data.get("avatar_decoration_data", None)
         self._public_flags = data.get("public_flags", 0)
         self.bot = data.get("bot", False)
         self.system = data.get("system", False)
+        return self
 
     @classmethod
     def _copy(cls: type[BU], user: BU) -> BU:
@@ -157,47 +138,14 @@ class BaseUser(_UserTag):
         return self
 
     def _to_minimal_user_json(self) -> dict[str, Any]:
-        return {
-            "username": self.name,
-            "id": self.id,
-            "avatar": self._avatar,
-            "discriminator": self.discriminator,
-            "global_name": self.global_name,
-            "bot": self.bot,
-        }
-
-    @property
-    def jump_url(self) -> str:
-        """Returns a URL that allows the client to jump to the user.
-
-        .. versionadded:: 2.0
-        """
-        return f"https://discord.com/users/{self.id}"
+        data = super()._to_minimal_user_json()
+        data["bot"] = self.bot
+        return data
 
     @property
     def public_flags(self) -> PublicUserFlags:
         """The publicly available flags the user has."""
         return PublicUserFlags._from_value(self._public_flags)
-
-    @property
-    def avatar(self) -> Asset | None:
-        """Returns an :class:`Asset` for the avatar the user has.
-
-        If the user does not have a traditional avatar, ``None`` is returned.
-        If you want the avatar that a user has displayed, consider :attr:`display_avatar`.
-        """
-        if self._avatar is not None:
-            return Asset._from_avatar(self._state, self.id, self._avatar)
-        return None
-
-    @property
-    def default_avatar(self) -> Asset:
-        """Returns the default avatar for a given user.
-        This is calculated by the user's ID if they're on the new username system, otherwise their discriminator.
-        """
-        eq = (self.id >> 22) if self.is_migrated else int(self.discriminator)
-        perc = 6 if self.is_migrated else 5
-        return Asset._from_default_avatar(self._state, eq % perc)
 
     @property
     def display_avatar(self) -> Asset:
@@ -280,50 +228,6 @@ class BaseUser(_UserTag):
         """
         return self.colour
 
-    @property
-    def mention(self) -> str:
-        """Returns a string that allows you to mention the given user."""
-        return f"<@{self.id}>"
-
-    @property
-    def created_at(self) -> datetime:
-        """Returns the user's creation time in UTC.
-
-        This is when the user's Discord account was created.
-        """
-        return snowflake_time(self.id)
-
-    @property
-    def display_name(self) -> str:
-        """Returns the user's display name.
-        This will be their global name if set, otherwise their username.
-        """
-        return self.global_name or self.name
-
-    def mentioned_in(self, message: Message) -> bool:
-        """Checks if the user is mentioned in the specified message.
-
-        Parameters
-        ----------
-        message: :class:`Message`
-            The message to check if you're mentioned in.
-
-        Returns
-        -------
-        :class:`bool`
-            Indicates if the user is mentioned in the message.
-        """
-
-        if message.mention_everyone:
-            return True
-
-        return any(user.id == self.id for user in message.mentions)
-
-    @property
-    def is_migrated(self) -> bool:
-        """Checks whether the user is already migrated to global name."""
-        return self.discriminator == "0"
-
 
 class ClientUser(BaseUser):
     """Represents your Discord user.
@@ -384,8 +288,8 @@ class ClientUser(BaseUser):
         mfa_enabled: bool
         _flags: int
 
-    def __init__(self, *, state: ConnectionState, data: UserPayload) -> None:
-        super().__init__(state=state, data=data)
+    def __init__(self, *, state: ConnectionState, id: int) -> None:
+        super().__init__(state=state, id=id)
 
     def __repr__(self) -> str:
         if self.is_migrated:
@@ -407,7 +311,7 @@ class ClientUser(BaseUser):
         )
 
     async def _update(self, data: UserPayload) -> None:
-        super()._update(data)
+        await super()._update(data)
         # There's actually an Optional[str] phone field as well, but I won't use it
         self.verified = data.get("verified", False)
         self.locale = data.get("locale")
@@ -479,10 +383,10 @@ class ClientUser(BaseUser):
             payload["banner"] = _bytes_to_base64_data(banner)
 
         data: UserPayload = await self._state.http.edit_profile(payload)
-        return ClientUser(state=self._state, data=data)
+        return await ClientUser.__load__(self._state, int(data["id"]), data)
 
 
-class User(BaseUser, discord.abc.Messageable):
+class User(BaseUser):
     """Represents a Discord user.
 
     .. container:: operations
@@ -527,8 +431,8 @@ class User(BaseUser, discord.abc.Messageable):
 
     __slots__ = ("_stored",)
 
-    def __init__(self, *, state: ConnectionState, data: UserPayload) -> None:
-        super().__init__(state=state, data=data)
+    def __init__(self, *, state: ConnectionState, id: int) -> None:
+        super().__init__(state=state, id=id)
         self._stored: bool = False
 
     def __repr__(self) -> str:
@@ -550,126 +454,3 @@ class User(BaseUser, discord.abc.Messageable):
         self = super()._copy(user)
         self._stored = False
         return self
-
-    async def _get_channel(self) -> DMChannel:
-        ch = await self.create_dm()
-        return ch
-
-    async def get_dm_channel(self) -> DMChannel | None:
-        """Returns the channel associated with this user if it exists.
-
-        If this returns ``None``, you can create a DM channel by calling the
-        :meth:`create_dm` coroutine function.
-        """
-        return await self._state._get_private_channel_by_user(self.id)
-
-    async def get_mutual_guilds(self) -> list[Guild]:
-        """The guilds that the user shares with the client.
-
-        .. note::
-
-            This will only return mutual guilds within the client's internal cache.
-
-        .. versionadded:: 1.7
-        """
-        return [
-            guild for guild in await self._state.cache.get_all_guilds() if await guild.get_member(self.id)
-        ]
-
-    async def create_dm(self) -> DMChannel:
-        """|coro|
-
-        Creates a :class:`DMChannel` with this user.
-
-        This should be rarely called, as this is done transparently for most
-        people.
-
-        Returns
-        -------
-        :class:`.DMChannel`
-            The channel that was created.
-        """
-        found = self.dm_channel
-        if found is not None:
-            return found
-
-        state = self._state
-        data: DMChannelPayload = await state.http.start_private_message(self.id)
-        return await state.add_dm_channel(data)
-
-    async def create_test_entitlement(self, sku: discord.abc.Snowflake) -> Entitlement:
-        """|coro|
-
-        Creates a test entitlement for the user.
-
-        Parameters
-        ----------
-        sku: :class:`Snowflake`
-            The SKU to create a test entitlement for.
-
-        Returns
-        -------
-        :class:`Entitlement`
-            The created entitlement.
-        """
-        payload = {
-            "sku_id": sku.id,
-            "owner_id": self.id,
-            "owner_type": 2,
-        }
-        data = await self._state.http.create_test_entitlement(self._state.application_id, payload)
-        return Entitlement(data=data, state=self._state)
-
-    def entitlements(
-        self,
-        skus: list[Snowflake] | None = None,
-        before: SnowflakeTime | None = None,
-        after: SnowflakeTime | None = None,
-        limit: int | None = 100,
-        exclude_ended: bool = False,
-    ) -> EntitlementIterator:
-        """Returns an :class:`.AsyncIterator` that enables fetching the user's entitlements.
-
-        This is identical to :meth:`Client.entitlements` with the ``user`` parameter.
-
-        .. versionadded:: 2.6
-
-        Parameters
-        ----------
-        skus: list[:class:`.abc.Snowflake`] | None
-            Limit the fetched entitlements to entitlements that are for these SKUs.
-        before: :class:`.abc.Snowflake` | :class:`datetime.datetime` | None
-            Retrieves guilds before this date or object.
-            If a datetime is provided, it is recommended to use a UTC-aware datetime.
-            If the datetime is naive, it is assumed to be local time.
-        after: :class:`.abc.Snowflake` | :class:`datetime.datetime` | None
-            Retrieve guilds after this date or object.
-            If a datetime is provided, it is recommended to use a UTC-aware datetime.
-            If the datetime is naive, it is assumed to be local time.
-        limit: Optional[:class:`int`]
-            The number of entitlements to retrieve.
-            If ``None``, retrieves every entitlement, which may be slow.
-            Defaults to ``100``.
-        exclude_ended: :class:`bool`
-            Whether to limit the fetched entitlements to those that have not ended.
-            Defaults to ``False``.
-
-        Yields
-        ------
-        :class:`.Entitlement`
-            The application's entitlements.
-
-        Raises
-        ------
-        :exc:`HTTPException`
-            Retrieving the entitlements failed.
-        """
-        return EntitlementIterator(
-            self._state,
-            sku_ids=[sku.id for sku in skus] if skus else None,
-            before=before,
-            after=after,
-            limit=limit,
-            user_id=self.id,
-            exclude_ended=exclude_ended,
-        )
